fix: wikipedia (zh) API error
